media: intel/ipu4: HACK: Re-purpose ipu6 files for ipu4 driver
This is of course not intended for upstream, but allows us to easily see
what the differences are between the ipu4 and ipu6 drivers.

This should in turn make it easier to make a decision of how much of the
drivers that should be common code, and how much should be separate
implementations.

Signed-off-by: Andreas Helbech Kleist <andreaskleist@gmail.com>