EMB-2014: media: intel/ipu4: Remove IPU6 psys_ipdata
At the early stages of the IPU4 driver port, the PCI driver module
intel-ipu6.ko was still meant to be compatible with both IPU4 and IPU6.
This was however down-prioritized in order to get something working
faster.

So now we remove the dead code for IPU6 in order to get a cleaner IPU4
driver.

At some future stage, it should be possible to create a shared PCI
driver, and then only have duplicate isys drivers.

Signed-off-by: Andreas Helbech Kleist <andreaskleist@gmail.com>