EMB-2177: media: intel/ipu4: Disable sof_event_by_stream in csi2_isr
The function ipu6_isys_csi2_sof_event_by_stream increments the sequence counter.
This function was called from both the CSI2 SOF ISR and when
IPU6_FW_ISYS_RESP_TYPE_FRAME_SOF was received.

It might be correct when using virtual channels, so I don't think this
should necessarily be fixed for the upstream driver, but for our use case it
is wrong.