media: intel/ipu4: video_release: Use vb2_fop_release to clear owner and queue
Without this we get

    Unable to request buffers: Device or resource busy (16).

when yavta calls ioctl VIDIOC_REQBUFS if the device has already been used once.

The difference between v4l2_fh_release and vb2_fop_release is that
vb2_fop_release also clears the queue. If the queue is not cleared when
VIDIOC_REQBUFS is called, we get the error mentioned above.

Signed-off-by: Andreas Helbech Kleist <andreaskleist@gmail.com>