media: intel/ipu4: Implement silent_reset when 'need_reset'
This solves the issue where the firmware is otherwise not able to open a
new stream (which is logged as 'isys power cycle required').

The solution is to pause the current stream, waiting for the firmware to
release all buffers. When that is done, the firmware stream is stopped,
the IPU4 isys is reset, and the firmware stream is re-opened.

The above all happens without userspace (gstreamer in our case) getting
notified about it. The only visible difference is that some frames might
be lost in this process.

Signed-off-by: Andreas Helbech Kleist <andreaskleist@gmail.com>